gcp-projects: port to new composability API
