+++ plugins/gcp-projects/package.json
@@ -37,7 +37,7 @@
     "@material-ui/lab": "4.0.0-alpha.45",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.0.0-beta.0",
     "react-use": "^15.3.3"
   },
   "devDependencies": {
